Revert "конфликты мб всё щас сломается нахуй и пизда"
This reverts commit d4572b250166cfbe184120b83c7dc196e4c87e45.

+++ code/game/gamemodes/shadowling/special_shadowling_abilities.dm
@@ -1,43 +1,30 @@
 //In here: Hatch and Ascendance
 GLOBAL_LIST_INIT(possibleShadowlingNames, list("U'ruan", "Y`shej", "Nex", "Hel-uae", "Noaey'gief", "Mii`mahza", "Amerziox", "Gyrg-mylin", "Kanet'pruunance", "Vigistaezian")) //Unpronouncable 2: electric boogalo)
 
-/datum/action/cooldown/spell/shadowling_hatch
+/obj/effect/proc_holder/spell/shadowling_hatch
 	name = "Hatch"
 	desc = "Сбрасывает вашу маскировку."
-	cooldown_time = 5 MINUTES
-	spell_requirements = SPELL_REQUIRES_NO_ANTIMAGIC | SPELL_REQUIRES_HUMAN
-	school = SCHOOL_FORBIDDEN
-	button_icon_state = "hatch"
-	background_icon_state = "bg_shadowling"
+	base_cooldown = 5 MINUTES
+	clothes_req = FALSE
+	action_icon_state = "hatch"
+	action_background_icon_state = "bg_shadowling"
 	var/cycles_unused = 0
-	var/hatching = FALSE
-	var/list/datum/action/cooldown/spell/shadowling_spells = list(
-		/datum/action/cooldown/spell/pointed/shadowling_enthrall,
-		/datum/action/cooldown/spell/aoe/shadowling_glare,
-		/datum/action/cooldown/spell/aoe/shadowling_veil,
-		/datum/action/cooldown/spell/jaunt/ethereal_jaunt/shadowling_shadow_walk,
-		/datum/action/cooldown/spell/aoe/shadowling_icy_veins,
-		/datum/action/cooldown/spell/shadowling_regen_armor,
-	)
 
-/datum/action/cooldown/spell/shadowling_hatch/can_cast_spell(feedback)
-	return ..() && !hatching
+/obj/effect/proc_holder/spell/shadowling_hatch/create_new_targeting()
+	return new /datum/spell_targeting/self
 
-/datum/action/cooldown/spell/shadowling_hatch/cast(atom/cast_on)
-	. = ..()
-	var/mob/living/carbon/human/user = cast_on
-	if(user.stat || !ishuman(user) || !user || !is_shadow(user) || user.isinspace())
+/obj/effect/proc_holder/spell/shadowling_hatch/cast(list/targets, mob/living/carbon/human/user = usr)
+	if(user.stat || !ishuman(user) || !user || !is_shadow(user) || isinspace(user))
 		return
 
 	if(!isturf(user.loc))
-		reset_spell_cooldown()
+		revert_cast(user)
 		to_chat(user, span_warning("Вы должны стоять на полу, чтобы раскрыться!"))
 		return
-	hatching = TRUE
+
 	if(tgui_alert(user, "Вы уверены, что хотите раскрыться? Вы не сможете прервать это!", "Hatch", list("Yes", "No")) != "Yes")
 		to_chat(user, span_warning("Вы решили не раскрываться сейчас."))
-		hatching = FALSE
-		reset_spell_cooldown()
+		revert_cast(user)
 		return
 
 	ADD_TRAIT(user, TRAIT_NO_TRANSFORM, UNIQUE_TRAIT_SOURCE(src))
@@ -138,40 +125,36 @@ GLOBAL_LIST_INIT(possibleShadowlingNames, list("U'ruan", "Y`shej", "Nex", "Hel-u
 	to_chat(user, span_shadowling("<b><i>Ваши силы пробудились. Теперь вы заживёте в полную меру. Помните свои цели. Сотрудничайте со своими союзниками и рабами.</b></i>"))
 	user.ExtinguishMob()
 	user.set_nutrition(NUTRITION_LEVEL_FED)
-	for(var/spell in shadowling_spells)
-		user.mind.AddSpell(new spell)
+	//user.mind.AddSpell(new /obj/effect/proc_holder/spell/shadowling_vision(null))
+	user.mind.AddSpell(new /obj/effect/proc_holder/spell/shadowling_enthrall(null))
+	user.mind.AddSpell(new /obj/effect/proc_holder/spell/shadowling_glare(null))
+	user.mind.AddSpell(new /obj/effect/proc_holder/spell/aoe/shadowling_veil(null))
+	user.mind.AddSpell(new /obj/effect/proc_holder/spell/shadowling_shadow_walk(null))
+	user.mind.AddSpell(new /obj/effect/proc_holder/spell/aoe/shadowling_icy_veins(null))
+	user.mind.AddSpell(new /obj/effect/proc_holder/spell/shadowling_regen_armor(null))
 
 	QDEL_NULL(user.hud_used)
 	user.set_hud_used(new /datum/hud/human(user, ui_style2icon(user.client.prefs.UI_style), user.client.prefs.UI_style_color, user.client.prefs.UI_style_alpha))
 	user.hud_used.show_hud(user.hud_used.hud_version)
-	user.RemoveSpell(src)
 
-/datum/action/cooldown/spell/shadowling_ascend
+/obj/effect/proc_holder/spell/shadowling_ascend
 	name = "Ascend"
 	desc = "Завершить свою истинную форму."
-	cooldown_time = 5 MINUTES
-	spell_requirements = SPELL_REQUIRES_HUMAN
-	school = SCHOOL_FORBIDDEN
-	button_icon_state = "ascend"
-	background_icon_state = "bg_shadowling"
-	var/list/ascendant_spells = list(
-		/datum/action/cooldown/spell/pointed/ascendant_annihilate,
-		/datum/action/cooldown/spell/pointed/ascendant_hypnosis,
-		/datum/action/cooldown/spell/jaunt/ascendant_phase_shift,
-		/datum/action/cooldown/spell/aoe/ascendant_storm,
-		/datum/action/cooldown/spell/ascendant_transmit,
-		/datum/action/cooldown/spell/pointed/shadowling_revive_thrall/ascendant,
-	)
+	base_cooldown = 5 MINUTES
+	clothes_req = FALSE
+	action_icon_state = "ascend"
+	action_background_icon_state = "bg_shadowling"
 
-/datum/action/cooldown/spell/shadowling_ascend/cast(atom/cast_on)
-	. = ..()
-	var/mob/living/carbon/human/user = cast_on
+/obj/effect/proc_holder/spell/shadowling_ascend/create_new_targeting()
+	return new /datum/spell_targeting/self
+
+/obj/effect/proc_holder/spell/shadowling_ascend/cast(list/targets, mob/living/carbon/human/user = usr)
 	if(!shadowling_check(user))
 		return
 
 	if(tgui_alert(user, "Время завершить свою форму. Вы уверены?", "Ascend", list("Yes", "No")) != "Yes")
 		to_chat(user, span_warning("Вы передумали завершать свою форму сейчас."))
-		reset_spell_cooldown()
+		revert_cast(user)
 		return
 
 	ADD_TRAIT(user, TRAIT_NO_TRANSFORM, PERMANENT_TRANSFORMATION_TRAIT)
@@ -224,29 +207,38 @@ GLOBAL_LIST_INIT(possibleShadowlingNames, list("U'ruan", "Y`shej", "Nex", "Hel-u
 
 	var/mob/living/simple_animal/ascendant_shadowling/ascendant = new (user.loc)
 	ascendant.announce("VYSHA NERADA YEKHEZET U'RUU!!", 5, 'sound/hallucinations/veryfar_noise.ogg')
-	for(var/datum/action/cooldown/spell/spell as anything in user.mind.spell_list)
-		if(spell == src || !spell.shadowling_spell)
+	for(var/obj/effect/proc_holder/spell/spell as anything in user.mind.spell_list)
+		if(spell == src)
 			continue
 		user.mind.RemoveSpell(spell)
 
 	user.mind.transfer_to(ascendant)
 	ascendant.name = user.real_name
 	ascendant.languages = user.languages
-	for(var/spell_to_add in ascendant_spells)
-		ascendant.mind.AddSpell(new spell_to_add)
+	ascendant.mind.AddSpell(new /obj/effect/proc_holder/spell/ascendant_annihilate(null))
+	ascendant.mind.AddSpell(new /obj/effect/proc_holder/spell/ascendant_hypnosis(null))
+	ascendant.mind.AddSpell(new /obj/effect/proc_holder/spell/ascendant_phase_shift(null))
+	ascendant.mind.AddSpell(new /obj/effect/proc_holder/spell/aoe/ascendant_storm(null))
+	ascendant.mind.AddSpell(new /obj/effect/proc_holder/spell/ascendant_transmit(null))
+	ascendant.mind.AddSpell(new /obj/effect/proc_holder/spell/shadowling_revive_thrall/ascendant(null))
 
 	if(ascendant.real_name)
 		ascendant.real_name = user.real_name
 
-	qdel(user)
+	user.invisibility = INVISIBILITY_OBSERVER	//This is pretty bad, but is also necessary for the shuttle call to function properly
+	user.forceMove(ascendant)
 
 	sleep(5 SECONDS)
+	if(QDELETED(user))
+		return
 
 	if(!SSticker.mode.shadowling_ascended)
 		sleep(60 SECONDS)
 		SSticker?.mode?.end_game()
 
 	SSticker.mode.shadowling_ascended = TRUE
+	ascendant.mind.RemoveSpell(src)
+	qdel(user)
 
 /**
  * Testing purpose.
@@ -270,16 +262,17 @@ GLOBAL_LIST_INIT(possibleShadowlingNames, list("U'ruan", "Y`shej", "Nex", "Hel-u
 
 	ExtinguishMob()
 	set_nutrition(NUTRITION_LEVEL_FED)
-	mind.AddSpell(new /datum/action/cooldown/spell/pointed/shadowling_enthrall)
-	mind.AddSpell(new /datum/action/cooldown/spell/aoe/shadowling_glare)
-	mind.AddSpell(new /datum/action/cooldown/spell/aoe/shadowling_veil)
-	mind.AddSpell(new /datum/action/cooldown/spell/jaunt/ethereal_jaunt/shadowling_shadow_walk)
-	mind.AddSpell(new /datum/action/cooldown/spell/aoe/shadowling_icy_veins)
-	mind.AddSpell(new /datum/action/cooldown/spell/shadowling_regen_armor)
-	mind.AddSpell(new /datum/action/cooldown/spell/aoe/shadowling_screech)
-	mind.AddSpell(new /datum/action/cooldown/spell/shadowling_blindness_smoke/)
-	mind.AddSpell(new /datum/action/cooldown/spell/pointed/shadowling_revive_thrall)
-	mind.AddSpell(new /datum/action/cooldown/spell/shadowling_ascend)
+	mind.AddSpell(new /obj/effect/proc_holder/spell/shadowling_enthrall(null))
+	mind.AddSpell(new /obj/effect/proc_holder/spell/shadowling_glare(null))
+	mind.AddSpell(new /obj/effect/proc_holder/spell/aoe/shadowling_veil(null))
+	mind.AddSpell(new /obj/effect/proc_holder/spell/shadowling_shadow_walk(null))
+	mind.AddSpell(new /obj/effect/proc_holder/spell/aoe/shadowling_icy_veins(null))
+	mind.AddSpell(new /obj/effect/proc_holder/spell/shadowling_regen_armor(null))
+	mind.AddSpell(new /obj/effect/proc_holder/spell/aoe/shadowling_screech(null))
+	mind.AddSpell(new /obj/effect/proc_holder/spell/shadowling_blindness_smoke(null))
+	mind.AddSpell(new /obj/effect/proc_holder/spell/shadowling_null_charge(null))
+	mind.AddSpell(new /obj/effect/proc_holder/spell/shadowling_revive_thrall(null))
+	mind.AddSpell(new /obj/effect/proc_holder/spell/shadowling_ascend(null))
 
 	mind.special_role = SPECIAL_ROLE_SHADOWLING
 	SSticker.mode.shadows += mind
